[KOGITO-4889] Renaming integration tests
