+++ integration-tests/integration-tests-quarkus-decisions/src/test/java/org/kie/kogito/integrationtests/quarkus/ElementAtIndexIT.java
@@ -23,55 +23,49 @@ import io.restassured.http.ContentType;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.nullValue;
 
 @QuarkusTest
-class BasicAddTest {
+class ElementAtIndexIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     }
 
     @Test
-    void testWholeModel() {
-        given().body("{ \"a\": \"v1\", \"b\": \"v2\" }")
+    void testHeaderEmpty() {
+        given().body(" { \"a list\" : [\"a\", \"b\", \"c\"], \"an index\" : 1 }")
                 .contentType(ContentType.JSON)
                 .when()
-                .post("/basicAdd")
+                .post("/ElementAtIndex")
                 .then()
                 .statusCode(200)
-                .body("decision", is("v1v2"));
+                .header("X-Kogito-decision-messages", nullValue())
+                .body("'element at index'", is("a"));
     }
 
     @Test
-    void testWholeModel_dmnresult() {
-        given().body("{ \"a\": \"v1\", \"b\": \"v2\" }")
+    void testHeaderPopulated() {
+        given().body(" { \"a list\" : [\"a\", \"b\", \"c\"], \"an index\" : 47 }")
                 .contentType(ContentType.JSON)
                 .when()
-                .post("/basicAdd/dmnresult")
+                .post("/ElementAtIndex")
                 .then()
                 .statusCode(200)
-                .body("dmnContext.decision", is("v1v2"));
+                .header("X-Kogito-decision-messages", notNullValue()) // a warning
+                .body("'element at index'", nullValue());
     }
 
     @Test
-    void testDs1() {
-        given().body("{ \"a\": \"v1\", \"b\": \"v2\" }")
-                .contentType(ContentType.JSON)
+    void testGET() {
+        given().accept(ContentType.XML)
                 .when()
-                .post("/basicAdd/ds1")
+                .get("/ElementAtIndex")
                 .then()
                 .statusCode(200)
-                .body(is("\"v1v2\"")); // a JSON string literal: "v1v2"
+                .body("definitions.decision[0].children().size()", is(4))
+                .body("definitions.children().findAll { node -> node.name() == 'literalExpression' }.size()", is(0));
     }
 
-    @Test
-    void testDs1_dmnresult() {
-        given().body("{ \"a\": \"v1\", \"b\": \"v2\" }")
-                .contentType(ContentType.JSON)
-                .when()
-                .post("/basicAdd/ds1/dmnresult")
-                .then()
-                .statusCode(200)
-                .body("dmnContext.decision", is("v1v2"));
-    }
 }
