+++ integration-tests/integration-tests-quarkus-decisions/src/test/java/org/kie/kogito/integrationtests/quarkus/OASIT.java
@@ -33,7 +33,7 @@ import static org.hamcrest.Matchers.aMapWithSize;
 import static org.hamcrest.Matchers.greaterThan;
 
 @QuarkusTest
-class OASTest {
+class OASIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
