+++ integration-tests/integration-tests-quarkus-decisions/src/test/java/org/kie/kogito/integrationtests/quarkus/OneOfEachTypeIT.java
@@ -24,7 +24,7 @@ import io.restassured.http.ContentType;
 import static org.hamcrest.CoreMatchers.is;
 
 @QuarkusTest
-class OneOfEachTypeTest {
+class OneOfEachTypeIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
