+++ integration-tests/integration-tests-quarkus-predictions/src/test/java/org/kie/kogito/integrationtests/quarkus/NativePMMLTreeIT.java
@@ -18,6 +18,6 @@ package org.kie.kogito.integrationtests.quarkus;
 import io.quarkus.test.junit.NativeImageTest;
 
 @NativeImageTest
-class NativePMMLTreeIT extends PMMLTreeTest {
+class NativePMMLTreeIT extends PMMLTreeIT {
 
 }
