+++ integration-tests/integration-tests-quarkus-processes/pom.xml
@@ -163,7 +163,7 @@
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
+        <artifactId>maven-failsafe-plugin</artifactId>
         <configuration>
           <includes>
             <include>org/kie/kogito/integrationtests/quarkus/*</include>
