+++ integration-tests/integration-tests-quarkus-processes/src/test/java/org/kie/kogito/integrationtests/quarkus/EnumsIT.java
@@ -38,7 +38,7 @@ import static org.hamcrest.Matchers.emptyOrNullString;
 
 @QuarkusTest
 @QuarkusTestResource(InfinispanQuarkusTestResource.Conditional.class)
-class EnumsTest {
+class EnumsIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
