+++ integration-tests/integration-tests-quarkus-processes/src/test/java/org/kie/kogito/integrationtests/quarkus/PingPongMessageIT.java
@@ -32,7 +32,7 @@ import static org.hamcrest.CoreMatchers.equalTo;
 
 @QuarkusTest
 @QuarkusTestResource(KafkaQuarkusTestResource.class)
-public class PingPongMessageTest {
+public class PingPongMessageIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
