+++ integration-tests/integration-tests-quarkus-processes/src/test/java/org/kie/kogito/integrationtests/quarkus/ProcessSvgAddonIT.java
@@ -37,7 +37,7 @@ import static org.hamcrest.CoreMatchers.equalTo;
 @QuarkusTest
 @QuarkusTestResource(InfinispanQuarkusTestResource.Conditional.class)
 @QuarkusTestResource(DataIndexWiremock.class)
-public class ProcessSvgAddonTest {
+public class ProcessSvgAddonIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
