+++ integration-tests/integration-tests-quarkus-processes/src/test/java/org/kie/kogito/integrationtests/quarkus/TaskIT.java
@@ -44,7 +44,7 @@ import static org.junit.Assert.assertEquals;
 
 @QuarkusTest
 @QuarkusTestResource(InfinispanQuarkusTestResource.Conditional.class)
-class TaskTest {
+class TaskIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
