+++ integration-tests/integration-tests-quarkus-processes/src/test/java/org/kie/kogito/integrationtests/quarkus/infinispan/InfinispanIT.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.integrationtests.quarkus;
+package org.kie.kogito.integrationtests.quarkus.infinispan;
 
-import java.util.HashMap;
-import java.util.Map;
+import javax.inject.Inject;
 
+import org.infinispan.client.hotrod.RemoteCacheManager;
 import org.junit.jupiter.api.Test;
 import org.kie.kogito.testcontainers.quarkus.InfinispanQuarkusTestResource;
 
@@ -28,58 +28,46 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.equalTo;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.Matchers.emptyOrNullString;
+import static org.hamcrest.Matchers.nullValue;
 
 @QuarkusTest
-@QuarkusTestResource(InfinispanQuarkusTestResource.Conditional.class)
-class FlexibleProcessTest {
+@QuarkusTestResource(InfinispanQuarkusTestResource.class)
+class InfinispanIT {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     }
 
-    @Test
-    void testInstantiateProcess() {
-        Map<String, String> params = new HashMap<>();
-        params.put("var1", "first");
-        params.put("var2", "second");
+    @Inject
+    RemoteCacheManager cacheManager;
 
-        String pid = given()
-                .contentType(ContentType.JSON)
+    @Test
+    void testPersistence() {
+        String processId = "greetings";
+        String cacheName = processId + "_store";
+
+        String pid = given().contentType(ContentType.JSON)
                 .when()
-                .body(params)
-                .post("/AdHocProcess")
+                .post("/greetings")
                 .then()
                 .statusCode(201)
-                .header("Location", not(emptyOrNullString()))
                 .body("id", not(emptyOrNullString()))
-                .body("var1", equalTo("Hello first! Script"))
-                .body("var2", equalTo("second Script 2"))
+                .body("test", nullValue())
                 .extract()
                 .path("id");
 
-        given()
-                .contentType(ContentType.JSON)
+        assertThat(cacheManager.getCacheNames()).contains(cacheName);
+        assertThat(cacheManager.getCache(cacheName).containsKey(pid)).isTrue();
+
+        given().contentType(ContentType.JSON)
                 .when()
-                .get("/AdHocProcess/{pid}", pid)
+                .delete("/management/processes/{processId}/instances/{processInstanceId}", processId, pid)
                 .then()
                 .statusCode(200);
-    }
 
-    @Test
-    void testProcessException() {
-        Map<String, String> params = new HashMap<>();
-        params.put("var1", "exception");
-        params.put("var2", "second");
-
-        given()
-                .contentType(ContentType.JSON)
-                .when()
-                .body(params)
-                .post("/AdHocProcess")
-                .then()
-                .statusCode(500);
+        assertThat(cacheManager.getCache(cacheName).containsKey(pid)).isFalse();
     }
 }
